vidconsole: Fix crash on kernel boot
This commit fixes a crash that occurred when booting the kernel with vidconsole
enabled due to recent changes.

Additionally, the boot logo is now displayed by default at boot. If an
interrupt (Ctrl + C) is detected, the uboot will switch to vidconsole mode.

This change also removes the use of `CONFIG_DRM_ROCKCHIP_SHOW_LOGO`.

Signed-off-by: Ajit Singh <ajeetsinghchahar2@gmail.com>